Struct piet_common::kurbo::RoundedRect

pub struct RoundedRect {
    rect: Rect,
    radius: f64,
}

A rectangle with equally rounded corners.

By construction the rounded rectangle will have non-negative dimensions and radius clamped to half size of the rect.

Fields
rect: Rect

Coordinates of the rectangle.

radius: f64

Radius of all four corners.

Methods

impl RoundedRect

pub fn new(x0: f64, y0: f64, x1: f64, y1: f64, radius: f64) -> RoundedRect

A new rectangle from minimum and maximum coordinates.

The result will have non-negative width, height and radius.

pub fn from_rect(rect: Rect, radius: f64) -> RoundedRect

A new rounded rectangle from a rectangle and corner radius.

The result will have non-negative width, height and radius.

pub fn from_points(p0: Point, p1: Point, radius: f64) -> RoundedRect

A new rectangle from two points.

The result will have non-negative width, height and radius.

pub fn from_origin_size(origin: Point, size: Vec2, radius: f64) -> RoundedRect

A new rectangle from origin and size.

The result will have non-negative width, height and radius.

pub fn width(&self) -> f64

The width of the rectangle.

pub fn height(&self) -> f64

The height of the rectangle.

pub fn radius(&self) -> f64

Radius of the rounded corners. #[inline]

pub fn rect(&self) -> Rect

The (non-rounded) rectangle.

pub fn origin(&self) -> Point

The origin of the rectangle.

This is the top left corner in a y-down space.

pub fn center(&self) -> Point

The center point of the rectangle.